# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse:
    properties:
      code:
        description: |-
          Code is a machine-readable status identifier for the client to act on.
          Use SCREAMING_SNAKE_CASE.
          Example: "CREATED" | "NOT_FOUND" | "VALIDATION_ERROR" | "INTERNAL_ERROR"
        type: string
      data:
        description: |-
          Data carries the operation's result payload.
          Omitted from the JSON output when nil.
      message:
        description: |-
          Message is a human-readable description of the outcome.
          On 4xx/5xx responses this field is masked with a reference ID before reaching the client.
          Example: "candidate created successfully"
        type: string
      success:
        description: |-
          Success indicates whether the operation completed without error.
          Example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    name: Nexus Team
  description: |-
    REST API for the Nexus AI-driven interview platform.

    Authentication: protected endpoints accept a JWT either as an `Authorization: Bearer {token}` header
    or as the HttpOnly `access_token` cookie set by the auth endpoints. In Swagger UI, click Authorize and enter `Bearer {token}`.

    Every response is wrapped in the standard envelope `{ success, message, code, data }`.
    On 4xx/5xx responses the `message` field may be masked with a reference ID.
  title: Nexus Interview API
  version: "1.0"
paths:
...
  /auth/logout:
    post:
      description: |-
        Invalidates the refresh token stored for the authenticated actor so it can no longer be used to mint
        new access tokens, then expires the `access_token` and `refresh_token` cookies in the browser.
        The current access token stays valid until its natural expiry.
      produces:
      - application/json
      responses:
        "200":
          description: logged out — cookies cleared
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "401":
          description: UNAUTHORIZED — missing or invalid access token
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Auth
  /auth/refresh:
    post:
      description: |-
        Reads the `refresh_token` HttpOnly cookie (scoped to /api/v1/auth), validates it against the token store,
        and issues a brand-new access/refresh pair. Both tokens are written back as HttpOnly cookies;
        the response body carries no token. If the refresh token is unknown or expired the cookies are cleared
        and the client must log in again.
      produces:
      - application/json
      responses:
        "200":
          description: tokens refreshed — new cookies set
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "401":
          description: UNAUTHORIZED — refresh cookie missing, invalid or expired
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
      summary: Refresh access token
      tags:
      - Auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the token as: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...